+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 - 2024-??-?? 4.1.18
     - deprecated RegionControlFeature
-    - XL Fixes
+    - XL fixes
         - fixed blank wavecal
+        - flag non-downloaded EEPROMs as "stubbed" to drive re-download attempts
+        - Marquee updates during download
 - 2024-11-25 4.1.17
     - Multichannel testing
         - added Ctrl-, and Ctrl-. to select previous and next spectrometer